# dmr_merged_one_folder/296_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DMR Poultries API",
    "description": "Backend API for Masters, Trips, Staff, and Operations modules.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "http://localhost:4000", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/operations/dashboard": {
      "get": {
        "tags": ["Operations Dashboard"],
        "summary": "Operations dashboard KPIs (Completed trips + Approved ops records)",
        "parameters": [
          {
            "name": "asOf",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "Reference date for today/week/month trip counts"
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard totals",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OperationsDashboard" }
              }
            }
          }
        }
      }
    },
    "/api/operations/trips": {
      "get": {
        "tags": ["Operations Trips"],
        "summary": "List trips with search and filters",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "fromDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "toDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "vehicleId", "in": "query", "schema": { "type": "integer" } },
          { "name": "supervisorId", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["Draft", "Pending", "Completed", "Deleted"]
            }
          },
          { "name": "includeDeleted", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "description": "Trip list" } }
      },
      "post": {
        "tags": ["Operations Trips"],
        "summary": "Create trip (draft or with start data)",
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/operations/trips/{id}": {
      "get": {
        "tags": ["Operations Trips"],
        "summary": "Get trip by id",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Trip" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Operations Trips"],
        "summary": "Update / autosave trip (steps 1–5 payload)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Updated trip" } }
      },
      "delete": {
        "tags": ["Operations Trips"],
        "summary": "Soft-delete trip (status=Deleted)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Soft deleted" } }
      }
    },
    "/api/operations/trips/{id}/steps/{step}": {
      "post": {
        "tags": ["Operations Trips"],
        "summary": "Submit a trip workflow step",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          {
            "name": "step",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["start", "farm", "pickup", "deliveries", "expenses"] }
          }
        ],
        "responses": { "200": { "description": "Step saved" } }
      }
    },
    "/api/operations/trips/{id}/status": {
      "patch": {
        "tags": ["Operations Trips"],
        "summary": "Trip status workflow: Draft → Pending → Completed | Deleted",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["Draft", "Pending", "Completed", "Deleted"]
                  },
                  "approvedBy": { "type": "string" },
                  "reason": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated" }, "400": { "description": "Invalid status" } }
      }
    },
    "/api/operations/trip-list": {
      "get": {
        "tags": ["Operations Trips"],
        "summary": "Trip List — read-only historical view of completed/approved trips only",
        "description": "Returns ONLY trips with status='Completed' AND deleted=false. Draft, Pending, Cancelled and Deleted trips are excluded by the database query itself — never by the client. Read-only: no POST/PUT/DELETE/PATCH is defined for this resource.",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Matches trip_no, vehicle_no, driver_name, supervisor_name, source_farm (case-insensitive)" },
          { "name": "fromDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "toDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "vehicleId", "in": "query", "schema": { "type": "integer" } },
          { "name": "supervisorId", "in": "query", "schema": { "type": "integer" } },
          { "name": "driverId", "in": "query", "schema": { "type": "integer" } },
          { "name": "farmId", "in": "query", "schema": { "type": "integer" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of completed/approved trips",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/TripSummary" } },
                    "meta": { "$ref": "#/components/schemas/PaginatedMeta" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/operations/trip-list/{id}": {
      "get": {
        "tags": ["Operations Trips"],
        "summary": "Trip List detail — full completed/approved trip",
        "description": "Returns the full hydrated trip ONLY when it is eligible (status='Completed' AND deleted=false). Draft, Pending and Deleted trips return 404.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Trip", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TripSummary" } } } },
          "404": { "description": "Not eligible / not found" }
        }
      }
    },
    "/api/operations/rate-entry": {
      "get": {
        "tags": ["Rate Entry"],
        "summary": "List trips eligible for rate entry (Approved/Completed, not deleted, not already locked)",
        "parameters": [
          { "name": "fromDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "toDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "vehicleNo", "in": "query", "schema": { "type": "string" } },
          { "name": "supervisorName", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "rateStatus", "in": "query", "schema": { "type": "string", "enum": ["Pending", "Entered"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Eligible trips" } }
      },
      "post": {
        "tags": ["Rate Entry"],
        "summary": "Save (upsert) a trip's rate entry — does not lock the trip. Rate must be ₹50-₹300.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tripId", "rate"],
                "properties": {
                  "tripId": { "type": "integer" },
                  "rate": { "type": "number", "minimum": 50, "maximum": 300 },
                  "deliveries": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["id", "rate"],
                      "properties": {
                        "id": { "type": "integer" },
                        "rate": { "type": "number", "minimum": 50, "maximum": 300 }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Rate entry saved" },
          "400": { "description": "Rate outside ₹50-₹300, or other validation error" },
          "409": { "description": "Trip not eligible, or rate already locked" }
        }
      }
    },
...
    "/api/operations/rate-entry/{id}": {
      "get": {
        "tags": ["Rate Entry"],
        "summary": "Get a rate entry by its own id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Rate entry" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Rate Entry"],
        "summary": "Edit a saved (unlocked) rate entry — rejected once locked. Rate must be ₹50-₹300.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Updated rate entry" },
          "400": { "description": "Rate outside ₹50-₹300" },
          "409": { "description": "Rate entry is locked" }
        }
      }
    },
...
    "/api/operations/shop-rates/{id}": {
      "get": { "tags": ["Shop Rates"], "summary": "Get shop rate", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } },
      "put": { "tags": ["Shop Rates"], "summary": "Update shop rate", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } },
      "delete": { "tags": ["Shop Rates"], "summary": "Soft-delete shop rate", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } }
    },
    "/api/operations/shop-rates/{id}/status": {
      "patch": {
        "tags": ["Shop Rates"],
        "summary": "Approve / reject / delete shop rate",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StatusPatch" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/operations/shop-sales/{id}": {
      "get": { "tags": ["Shop Sales"], "summary": "Get shop sale", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } },
      "put": { "tags": ["Shop Sales"], "summary": "Update shop sale", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } },
      "delete": { "tags": ["Shop Sales"], "summary": "Soft-delete shop sale", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } }
    },
    "/api/operations/shop-sales/{id}/status": {
      "patch": {
        "tags": ["Shop Sales"],
        "summary": "Approval workflow for shop sale",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StatusPatch" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "OpsRecordStatus": {
        "type": "string",
        "enum": ["Draft", "Pending Approval", "Approved", "Rejected", "Deleted"]
      },
      "StatusPatch": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "$ref": "#/components/schemas/OpsRecordStatus" },
          "approvedBy": { "type": "string" },
          "rejectedBy": { "type": "string" },
          "rejectedReason": { "type": "string" },
          "reason": { "type": "string" }
        }
      },
      "OperationsDashboard": {
        "type": "object",
        "properties": {
          "totalTrips": { "type": "number" },
          "totalWeight": { "type": "number" },
          "totalSales": { "type": "number" },
          "totalCollections": { "type": "number" },
          "pendingCollections": { "type": "number" },
          "fuelExpenses": { "type": "number" },
          "todaysTrips": { "type": "number" },
          "weeklyTrips": { "type": "number" },
          "monthlyTrips": { "type": "number" }
        }
      },
      "PaginatedMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
    }
  }
}